pyproject.toml
[project]
name = "controlflow"
dynamic = ["version"]
description = "A framework for building agentic LLM workflows"
authors = [
    { name = "Jeremiah Lowin", email = "153965+jlowin@users.noreply.github.com" },
]
dependencies = [
    "prefect>=3.0",
    "jinja2>=3.1.4",
    "langchain_core>=0.3",
    "langchain_openai>=0.2",
    "langchain-anthropic>=0.2",
    "markdownify>=0.12.1",
    "openai<1.47",              # 1.47.0 introduced a bug with attempting to reuse an async client that doesnt have an obvious solution
    "pydantic-settings>=2.2.1",
    "textual>=0.61.1",
    "tiktoken>=0.7.0",
    "typer>=0.10",
]
readme = "README.md"
requires-python = ">= 3.9"
keywords = [
    "ai",
    "chatbot",
    "llm",
    "ai orchestration",
    "llm orchestration",
    "agentic workflows",
    "flow engineering",
    "prefect",
    "workflow",
    "orchestration",
    "python",
    "GPT",
    "openai",
    "assistant",
    "agents",
    "AI agents",
    "natural language processing",
]

[project.urls]
Code = "https://github.com/PrefectHQ/ControlFlow"

[project.optional-dependencies]
tests = [
    "chromadb",
    "duckduckgo-search",
    "langchain_community",
    "langchain_google_genai",
    "langchain_groq",
    "pytest-asyncio>=0.18.2,!=0.22.0,<0.23.0",
    "pytest-env>=0.8,<2.0",
    "pytest-rerunfailures>=10,<14",
    "pytest-sugar>=0.9,<2.0",
    "pytest>=7.0",
    "pytest-timeout",
    "pytest-xdist",
]
dev = [
    "controlflow[tests]",
    "ipython",
    "pdbpp",
    "pre-commit",
    "ruff>=0.3.4",
    "textual-dev",
    "mypy",
]

[project.scripts]
controlflow = "controlflow.cli.main:app"

[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "src/controlflow/_version.py"

[tool.rye]
managed = true

# ruff configuration
[tool.ruff]
target-version = "py311"
lint.select = ["I"]                                               # Changed from lint.extend-select to select
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$" # default, but here in case we want to change it

[tool.ruff.format]
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ['I', 'F401', 'E402']
"conftest.py" = ["F401", "F403"]
'tests/fixtures/*.py' = ['F401', 'F403']
# "src/controlflow/utilities/types.py" = ['F401']

[tool.pytest.ini_options]
timeout = 120
asyncio_mode = "auto"
norecursedirs = [
    "*.egg-info",
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".vscode",
    "node_modules",
]
testpaths = ["tests"]
env = [
    "CONTROLFLOW_TEST_MODE=1",
    # use 4o-mini for tests by default
    'D:CONTROLFLOW_LLM_MODEL=openai/gpt-4o-mini',
    'D:CONTROLFLOW_TOOLS_VERBOSE=1',
    'D:CONTROLFLOW_PRETTY_PRINT_AGENT_EVENTS=0',
    'D:CONTROLFLOW_LOG_LEVEL=DEBUG',
    'D:PREFECT_LOGGING_LEVEL=DEBUG',
]
